+++ ogmios-datum-cache.cabal
@@ -66,7 +66,6 @@ library
     App.Types
     Block.Fetch
     Block.Types
-    Config
     Database
     Parameters
